fix: parse expiresIn as Long per api contract; map unknown driver status to UNKNOWN; refactor: centralize user error messages in AppError.toUserMessage; format wallet balance from integer cents

+++ app/src/main/java/com/master/transportes/driver/feature/auth/data/dto/LoginResponseDto.kt
@@ -4,5 +4,5 @@ data class LoginResponseDto(
     val accessToken: String,
     val refreshToken: String,
     val sessionId: String,
-    val expiresIn: Double
+    val expiresIn: Long
 )
